Made shapes take doublesinstead of ints

+++ cps/shape.cpp
@@ -4,6 +4,7 @@
 
 #include "shape.h"
 
+// Base Class
 double Shape::get_height() const
 {
     return _height;
@@ -24,13 +25,14 @@ void Shape::set_width(double width)
     _width = width;
 }
 
-void Circle::setRadius(int radius)
+// Circle Class
+void Circle::setRadius(double radius)
 {
     set_height(radius*2);
     set_width(radius*2);
 }
 
-Circle::Circle(int radius)
+Circle::Circle(double radius)
 {
     setRadius(radius);
 }
@@ -40,6 +42,7 @@ std::stringstream Circle::generate()
     return std::stringstream("0 0 " + std::to_string(get_height()/2) + " 0 360 arc stroke");
 }
 
+// Rectangle Class
 std::stringstream Rectangle::generate()
 {
     return std::stringstream("newpath\n"
@@ -51,7 +54,7 @@ std::stringstream Rectangle::generate()
                              "stroke");
 }
 
-Rectangle::Rectangle(int width, int height)
+Rectangle::Rectangle(double width, double height)
 {
     set_height(height);
     set_width(width);
